+++ codec/encode/speex/include/speex/speex_types.h
@@ -108,7 +108,7 @@
    typedef short spx_int16_t;
    typedef unsigned short spx_uint16_t;
    typedef long spx_int32_t;
-   typedef unsigned long spx_uint32_t;
+   typedef uint32_t spx_uint32_t;
 
 #elif defined(CONFIG_TI_C6X)
